Allow plotter/plot.py to operate directly on experiment outputs
The CSV file created by experiments contains a configuration header, and so is not a true CSV file throughout. Prior to this change, the config header needed to be manually split into a separate file with a .cfg ending, and a reduced .csv file without the config header needed to be made; then these needed to be placed into a dated directory, and the directory zipped into a data.zip. With this new change, the script can be run directly on the file (or several files) created by an experiment.