Fix magnetic field equations
The code contained two mistakes. Firstly, the methods for computing the field angles were incorrect. In Taylor et al. 2021 Bioinspir. Biomim., Eq. 8 is correct; the diagram in Fig. 3B shows the angle delta is measured from the y-axis and is therefore not equivalent to theta + lambda (measured from the x-axis).

This formula is correctly implemented in the plotting code, so the plots hid this error.

Secondly, the negated sensing parameter, eta, should only apply to beta, and not to gamma. Otherwise, it is not possible to get stable equilibrium points at the goals with the default field parameters.