Fix steering algorithm
The old calculation for delta was incorrect in many critical ways.

Chief among these was the lack of parentheses around the 2*pi, which caused the modulo to operate on the 2 alone. This resulted in the robot turning towards different targets depending on its initial orientation, in general not towards the desired angle.

Furthermore, the old method caused the turning speed to increase in magnitude as delta got smaller. This meant that as the robot neared the (erroneously computed) angle at which it would stop turning, it frequently overturned, resulting in an unusual oscillating rotation motion.

This fix computes delta correctly, so that the angle the robot ultimately turns towards is independent of its starting orientation, and the turn speed decreases as it nears the desired angle, preventing oversteering.