Fix acknowledgment of close command
Prior to this change, successful runs would terminate with both host and client throwing errors, complaining "The last command was not properly acknowledged".

The critical fixes provided here are in host.py, where (1) the wrong socket was being checked for a close command from the client, and (2) the rate of sending data needed to be throttled to allow time for the client to determine it had reached its final goal and send the close command. Without this delay, the host would move on to sending data again so quickly that the client's close command wouldn't be received by the host until the host was already awaiting acknowledgment of the latest sent data; the result was that both host and client complained of unacknowledged messages.

Additional changes here include stripping out the `_finished` flag from the client, which was never used. The host never sends a close command, so the client does not need a special way to handle this.