Add `secular_variation_strategy` config param
Currently accepted values are `'none'` and `'imprint'`. When `'imprint'`, the agent performs an extra step each time it reaches a goal: It imprints on the latest magnetic signature at the goal location, and remembers that for the next circuit.

The `'imprint'` strategy changes the targeted magnetic signatures each circuit, so to visualize this we plot the up-to-date, targeted magnetic signature with a cross. When imprinting is working well, this should hardly move from the goal location. We continue to plot the secular variation of the original magnetic signature (from the first circuit) using a thin line to help visualize the effect of the time-varying magnetic field.